feat(admin): open trip detail from recent trips list (#283)
- Extract TripMiniMap to a shared component (TripMiniMap.tsx) so both
  StatsPage and AdminPage can reuse it without duplicating 100 lines
- AdminPage: recent trip cards are now clickable buttons
- AdminPage: tapping a trip opens a read-only bottom sheet with the
  user's name, GPS map (if available), and distance/CO₂/money stats
- useTrip() fetches the full trip detail (including gpsPoints) on demand

Closes #283
